outsourced my model to Student.js and exported the default Student model
